Move entities mapping to the application layer

+++ MessengerAPI.Presentation/Schemas/Channels/CreateMessageRequestSchema.cs
@@ -3,6 +3,6 @@ namespace MessengerAPI.Presentation.Schemas.Channels;
 public record CreateMessageRequestSchema
 {
     public string Text { get; init; }
-    public int? ReplyTo { get; init; }
+    public long? ReplyTo { get; init; }
     public List<Guid>? Attachments { get; init; }
 }
